UPSTREAM: <carry>: Exclude false positives from gitleaks scanner
There are a ton of credentials in our test suite because of how keda
works, but every last one of them is fake/useless. This just excludes
our test files from the scanner so they will stop generating
false-positives.

The preferred method to deal with them is to comment the exact line as
"# notasecret" but that's not feasible given the size of the test suite
and the number of fake credentials.

This should prevent the scary false positive ALERT! emails.

Signed-off-by: John Kyros <jkyros@redhat.com>