UPSTREAM: <carry>: e2e: OCP 4.22 timing regressions for pause and prometheus
So before the 4.22 branch cut, we were passing reliably, and now we're
failing reliably, and it's definitely a time problem. Digging into it,
it looks like finalizations are taking 10 seconds to clean up a
namespace, and it's just _slightly_ slower to retrieve a namespace, so
when the tests go through all the namespaces, downstream we have like 80
of them and all of that adds up to us blowing out our time window.

This just waits a little bit after a namespace deletion to make sure it
actually gets finalized and removed so we don't collide with it, and
then adjusts the metrics tests to wait a little bit for metrics
reconciliation, and avoid going through _every_ namespaces for the cases
where we only really want a specific namespace's metrics.

Signed-off-by: John Kyros <jkyros@redhat.com>

So there we were, we thought we fixed it, and Max was like "hey should
we replace all the calls or just the ones that are failing", and I was
like "we'll figure that out later". Well, now because of how we're
running the tests, we have some bleed-through where it looks like the
cron test creates some things that the metrics tests pick up, so this
replaces all the calls to testOperatorMetricValues with the retry
equivalent so we get a chance for our metrics to settle in properly.

This also adds more wait time to to the pause scale in tests that we
added to the pause scaledobject tests that makes them wait longer for
replicas to stabilize on OpenShift.

Also, the pause scale in test wasn't waiting for the proper replica
count before testing, so we were catching it on the way in, this just
makes it wait for stabilization before testing.

Signed-off-by: John Kyros <jkyros@redhat.com>